move user related route function to users blueprint
